12 hour formart for time slots
