+++ src/components/schedule/TheSchedule.svelte
@@ -1,12 +1,12 @@
 <script lang="ts">
   import type { GoogleColor } from "../../data/data";
   import type { TimeSlot } from "../../data/rooms";
-  import { getAmPmFromHours } from "../utils";
+  import { calculateSlotData } from "../utils/timeSlotHelper";
 
-  export let timeSlots: TimeSlot[] | null;
+  export let timeSlots: TimeSlot[];
   export let freeColor: GoogleColor;
 
-  let calculatedSlots: { [key: string]: any }[] = [];
+  let calculatedSlots: { [key: string]: any }[];
 
   $: {
     if (timeSlots) {
@@ -31,38 +31,6 @@
         };
       });
     }
-  }
-
-  function calculateSlotData(timeSlot: TimeSlot) {
-    const isFreeSlot = timeSlot.courseName === "Free";
-    const startTime = `${timeSlot.timeStart.hour}:${
-      timeSlot.timeStart.minute === 0 ? "00" : timeSlot.timeStart.minute
-    } ${getAmPmFromHours(timeSlot.timeStart.hour)}`;
-    const endTime = `${timeSlot.timeEnd.hour}:${
-      timeSlot.timeEnd.minute === 0 ? "00" : timeSlot.timeEnd.minute
-    } ${getAmPmFromHours(timeSlot.timeEnd.hour)}`;
-    const minutesTotal =
-      (timeSlot.timeEnd.hour - timeSlot.timeStart.hour) * 60 +
-      timeSlot.timeEnd.minute -
-      timeSlot.timeStart.minute;
-
-    let scheduleOffset = 0;
-    if (timeSlot.timeStart.hour > 7 || timeSlot.timeStart.minute > 0) {
-      const hoursInMinutes = (timeSlot.timeStart.hour - 7) * 60;
-      const offsetMinutes = hoursInMinutes + timeSlot.timeStart.minute;
-      scheduleOffset = offsetMinutes * (4.5 / 60);
-    }
-
-    const minutesRemRatio = isFreeSlot ? 4.5 / 60 : 4.5 / (60 * 1.05);
-
-    return {
-      isFreeSlot,
-      startTime,
-      endTime,
-      minutesTotal,
-      scheduleOffset,
-      minutesRemRatio,
-    };
   }
 </script>
