--- src/components/utils.ts
@@ -1,9 +1,0 @@
-export function getAmPmFromHours(
-  hours: number,
-  isCapital: boolean = false
-): string {
-  const pm = isCapital ? "PM" : "pm";
-  const am = isCapital ? "AM" : "am";
-
-  return hours >= 12 ? pm : am;
-}
